+++ python/GafferCyclesTest/IECoreCyclesPreviewTest/RendererTest.py
@@ -582,6 +582,86 @@ class RendererTest( GafferTest.TestCase ) :
 		self.assertEqual( color.g, points["N"].data[0].y )
 		self.assertEqual( color.b, points["N"].data[0].z )
 
+	def __testMeshSmoothing( self, cube, smoothingExpected ) :
+
+		renderer = GafferScene.Private.IECoreScenePreview.Renderer.create(
+			"Cycles",
+			GafferScene.Private.IECoreScenePreview.Renderer.RenderType.Interactive,
+		)
+
+		renderer.output(
+			"meshSmoothing",
+			IECoreScene.Output(
+				"test",
+				"ieDisplay",
+				"rgba",
+				{
+					"driverType" : "ImageDisplayDriver",
+					"handle" : "meshSmoothing",
+				}
+			)
+		)
+
+		# Render the cube, rotated so an edge faces the camera, shaded with the
+		# standard facing-ratio shader.
+
+		cubeObject = renderer.object( "/cube", cube, renderer.attributes( IECore.CompoundObject() ) )
+		cubeObject.transform( imath.M44f().translate( imath.V3f( 0, 0, 2 ) ).rotate( imath.V3f( 0, math.pi / 4.0, 0 ) ) )
+
+		renderer.render()
+		time.sleep( 2 )
+
+		del cubeObject
+		del renderer
+
+		# Check the shading on the center edge, and close to the back left and right edges.
+		# If normals have been smoothed, then the back edges should be close to zero.
+
+		image = IECoreImage.ImageDisplayDriver.storedImage( "meshSmoothing" )
+		self.assertTrue( isinstance( image, IECoreImage.ImagePrimitive ) )
+
+		center = self.__colorAtUV( image, imath.V2f( 0.5, 0.5 ) )
+		left = self.__colorAtUV( image, imath.V2f( 0.3, 0.5 ) )
+		right = self.__colorAtUV( image, imath.V2f( 0.7, 0.5 ) )
+
+		# Everything should have solid alpha.
+
+		self.assertEqual( center[3], 1 )
+		self.assertEqual( left[3], 1 )
+		self.assertEqual( right[3], 1 )
+
+		# Shading is down to whether the normals are smoothed or not.
+
+		if smoothingExpected :
+			# Center normal faces straight at us
+			self.assertGreater( center[0], 0.95 )
+			# Outer normals actually face slightly away
+			self.assertLess( left[0], 0 )
+			self.assertLess( right[0], 0 )
+		else :
+			# Everything faces towards and to the side of us.
+			self.assertGreater( center[0], 0.4 )
+			self.assertGreater( left[0], 0.4 )
+			self.assertGreater( right[0], 0.4 )
+
+	def testNoMeshNormals( self ) :
+
+		cube = IECoreScene.MeshPrimitive.createBox( imath.Box3f( imath.V3f( -0.5 ), imath.V3f( 0.5 ) ) )
+		del cube["N"]
+		self.__testMeshSmoothing( cube, smoothingExpected = False )
+
+	def testFaceVaryingMeshNormals( self ) :
+
+		# These are treated like non-existent normals, since Cycles doesn't support them.
+		cube = IECoreScene.MeshPrimitive.createBox( imath.Box3f( imath.V3f( -0.5 ), imath.V3f( 0.5 ) ) )
+		self.__testMeshSmoothing( cube, smoothingExpected = False )
+
+	def testVertexMeshNormals( self ) :
+
+		cube = IECoreScene.MeshPrimitive.createBox( imath.Box3f( imath.V3f( -0.5 ), imath.V3f( 0.5 ) ) )
+		cube["N"] = IECoreScene.MeshAlgo.calculateVertexNormals( cube, IECoreScene.MeshAlgo.NormalWeighting.Equal )
+		self.__testMeshSmoothing( cube, smoothingExpected = True )
+
 	def testUnsupportedPrimitiveVariables( self ) :
 
 		renderer = GafferScene.Private.IECoreScenePreview.Renderer.create(